feat: modularize page header
